code-python/simple-gradio-app/requirements.txt
gradio
torch 
torchvision 
opencv-python
pillow